consumer changes (#47)
* Consumer Details working

* Edit Details layout added

* updating (#1)

* Consumer Details working (#11)

* User details and image change option added

* resolving merge problem

* updating (#2)

* resolving merge conflict

* merge conflict

* updating (#3)

* Updating and resolving merge conflict

* consumer details bugs fixed

* resolving merge conflict

* resolving merge conflict #4

* resolving merge conflict #5

* resolving merge conflict #6

* resolving merge conflict #7

* resolving merge conflict #8

* resolving merge conflict #9

* resolving merge conflict #9

* Update ConsumerDetailsFragment.java

* Update ConsumerDetailsFragment.java

* Update ConsumerDetailsFragment.java

* Update ConsumerDetailsFragment.java

* updating (#4)

*updating

* final commit

* working on mass orders

* Provider changes

+++ app/src/main/java/com/example/student/homemade/ConsumerDetailsClass.java
@@ -2,12 +2,12 @@ package com.example.student.homemade;
 
 public class ConsumerDetailsClass {
     String username, password, address, contactNumber, email,typeOfUser;
-    float wallet;
-
+    Double wallet;
     public ConsumerDetailsClass() {
     }
 
-    public ConsumerDetailsClass(String username, String password, String address, String contactNo, float wallet, String email, String typeOfUser) {
+    public ConsumerDetailsClass(String username, String password, String address, String contactNo, Double wallet, String email, String typeOfUser) {
+
         this.username = username;
         this.password = password;
         this.address = address;
@@ -49,11 +49,12 @@ public class ConsumerDetailsClass {
         this.contactNumber = contactNumber;
     }
 
-    public float getWallet() {
+    public Double getWallet() {
         return wallet;
     }
 
-    public void setWallet(float wallet) {
+    public void setWallet(Double wallet) {
+
         this.wallet = wallet;
     }
